ranking

+++ src/main/java/com/example/sa_project/Ranking/UserProgress.java
@@ -22,6 +22,7 @@ public class UserProgress {
     private User user;
 
 
+    //이거 자료형...
     @Column(name = "experiencePoint")
-    private int experiencePoint;
+    private Long experiencePoint;
 }
